A program similar to the popular Wordle word game
Implement a program that behaves similarly to the popular Wordle word game using string, array, loop, and command-line argument in C.

At the time the user executes the program, they should decide the length of the word they want to guess between 5 and 8 letters. User is allowed one more guess than the length of the word. User may get some clues along the way, using those clues to try to home in on the correct word.
If after user guesses that letter turns green, it means that letter is in the secret word and is in the correct position. 
If it turns yellow, it means that the letter guessed appears somewhere in the word, but not in that spot. 
Letters that turn red aren’t in the word at all and can be omitted from future guesses.