A program spell-checks words in a text file
Implement the fastest program to spell-check a text file using hash table (a type of data structure).

This program is designed to load a default dictionary of words from disk into memory. Then, it checks an input text file and reports each misspelling in that file along with a bunch of statistics about the execution time of the “load”, “check”,  “size”, and “unload” functions.