A program identifies a person based on their DNA
Implement a program that identifies a person based on their DNA using command-line arguments, CSV reader functions, list, and dictionary in Python.
It will take a sequence of DNA and a CSV file containing STR counts for a list of individuals and then output to whom the DNA belongs.

DNA, the carrier of genetic information in living things, is a sequence of molecules called nucleotides. Each nucleotide of DNA contains one of four different bases: adenine (A), cytosine (C), guanine (G), or thymine (T). One place where DNA tends to have high genetic diversity is in Short Tandem Repeats (STRs). The number of times any particular STR repeats varies a lot among individuals.

In this program, a DNA database will be created as a CSV file, wherein each row corresponds to an individual, and each column corresponds to a particular STR. Then it will look through the DNA sequence for the longest consecutive sequence of repeated AGATs, AATGs, and TATCs. Finally, from this evidence, it will print out the name to whom the DNA belongs.