Refactoring a Hogwarts database
Practice refactoring a SQL database to eliminate redundancies.
Use SQL SELECT, CREATE, and INSERT statements to reorganize data.
Write Python to load new SQL tables.

The keeper of the Hall of Prophecy, whose job entailed labelling records and keeping them up to date, decided to create a SQL database of Hogwarts students. Unfortunately the database was poorly designed! The database contains only one table, students, with the name and head of each of the four Hogwarts Houses.
A better design would contain a students table (for only students), a houses table (for only houses), and a table that codifies the relationship between students and houses. This process of changing the “schema” of a database is known as refactoring.