Solving a mystery in Fiftyville
Based on the information in the database, write SQL queries to solve a mystery in Fiftyville.

The CS50 Duck has been stolen! Authorities believe that the thief stole the duck and then, shortly afterwards, took a flight out of town with the help of an accomplice. The goal is to identify:
• Who the thief is,
• What city the thief escaped to, and
• Who the thief’s accomplice is who helped them escape
All you know is that the theft took place on July 28, 2021 and that it took place on Humphrey Street.

The Fiftyville authorities have taken some of the town’s records from around the time of the theft and prepared a SQLite database, fiftyville.db, which contains tables of data from around the town. Using SQL SELECT queries to access the data and find the answers for the above questions.